pyproject.toml
[project]
name = "appium-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "appium-python-client==3.2.1",
    "pytest==7.4.3",
    "pytest-html==4.1.1",
    "selenium==4.15.2",
]
